toolkit/components/passwordmgr/test/mochitest/test_autofill_hasBeenTypePassword.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test no autofill into a password field that is no longer type=password</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="pwmgr_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
Login Manager test: Test no autofill into a password field that is no longer type=password

<script>
let readyPromise = registerRunTests();

runInParent(function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");
  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"].createInstance(Ci.nsILoginInfo);
  login1.init("https://example.com", "https://autofill", null, "user1", "pass1");
  Services.logins.addLogin(login1);
});

/** Test for Login Manager: Test no autofill into a password field that is no longer type=password **/

add_task(async function setup() {
  ok(readyPromise, "check promise is available");
  await readyPromise;
});


// As a control, test that autofill is working on this page.
add_task(async function test_autofill_control() {
  let formProcessedPromise = promiseFormsProcessed();
  // eslint-disable-next-line no-unsanitized/property
  document.querySelector("#content").innerHTML = `
    <form id="form1" action="https://autofill">
      <p>This is form 1.</p>
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">

      <button type="submit">Submit</button>
    </form>`;
  await formProcessedPromise;
  checkLoginForm($_(1, "uname"), "user1", $_(1, "pword"), "pass1");
});

add_task(async function test_no_autofill() {
  let formProcessedPromise = promiseFormsProcessed();
  // eslint-disable-next-line no-unsanitized/property
  document.querySelector("#content").innerHTML = `
    <form id="form1" action="https://autofill">
      <p>This is form 1.</p>
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">

      <button type="submit">Submit</button>
    </form>`;
  // Synchronously change the password field type to text before the fill happens.
  $_(1, "pword").type = "text";
  await formProcessedPromise;
  checkLoginForm($_(1, "uname"), "", $_(1, "pword"), "");
});
</script>

<p id="display"></p>

<div id="content"></div>

<pre id="test"></pre>
</body>
</html>
